+++ src/lib/libast/features/signal.c
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -311,7 +311,7 @@ main(void)
 
 	k = 0;
 	for (i = 0; map[i].name; i++)
-		if ((j = map[i].value) > 0 && j < elementsof(mapindex) && !mapindex[j])
+		if ((j = map[i].value) > 0 && j < (ssize_t)elementsof(mapindex) && !mapindex[j])
 		{
 			if (j > k)
 				k = j;
@@ -325,9 +325,9 @@ main(void)
 #else
 	j = i;
 #endif
-	if (j >= elementsof(mapindex))
-		j = elementsof(mapindex) - 1;
-	if (i <= j && i > 0 && i < elementsof(mapindex) && j > 0 && j < elementsof(mapindex))
+	if (j >= (ssize_t)elementsof(mapindex))
+		j = (ssize_t)elementsof(mapindex) - 1;
+	if (i <= j && i > 0 && i < (ssize_t)elementsof(mapindex) && j > 0 && j < (ssize_t)elementsof(mapindex))
 	{
 		if (j > k)
 			k = j;
